Refactored view to provide OpenSlides JavaScript files. All JavaScript files are concated now.
