+++ gulpfile.js
@@ -1,4 +1,4 @@
-/*!
+/**
  * Gulp tasks for development and production.
  *
  * Run
@@ -17,6 +17,7 @@ var argv = require('yargs').argv,
     concat = require('gulp-concat'),
     gulpif = require('gulp-if'),
     gettext = require('gulp-angular-gettext'),
+    jshint = require('gulp-jshint'),
     mainBowerFiles = require('main-bower-files'),
     minifyCSS = require('gulp-minify-css'),
     path = require('path'),
@@ -25,9 +26,14 @@ var argv = require('yargs').argv,
 // Directory where the results go
 var output_directory = path.join('openslides', 'static');
 
+
+/**
+ * Default tasks to be run before start.
+ */
+
 // Catches all JavaScript files from all bower components and concats them to
 // one file js/openslides-libs.js. In production mode the file is uglified.
-gulp.task('js', function() {
+gulp.task('js-libs', function() {
     return gulp.src(mainBowerFiles({
             filter: /\.js$/
         }))
@@ -38,7 +44,7 @@ gulp.task('js', function() {
 
 // Catches all CSS files from all bower components and concats them to one file
 // css/openslides-libs.css. In production mode the file is uglified.
-gulp.task('css', function() {
+gulp.task('css-libs', function() {
     return gulp.src(mainBowerFiles({
             filter: /\.css$/
         }))
@@ -48,12 +54,21 @@ gulp.task('css', function() {
 });
 
 // Catches all font files from all bower components.
-gulp.task('fonts', function() {
+gulp.task('fonts-libs', function() {
     return gulp.src(mainBowerFiles({
             filter: /\.(eot)|(svg)|(ttf)|(woff)|(woff2)$/
         }))
         .pipe(gulp.dest(path.join(output_directory, 'fonts')));
 });
+
+// Gulp default task. Runs all other tasks before.
+gulp.task('default', ['js-libs', 'css-libs', 'fonts-libs'], function() {});
+
+
+/**
+ * Extra tasks that have to be called manually. Useful for development and
+ * packageing.
+ */
 
 // Extracts translatable strings using angular-gettext and saves them in file
 // openslides/locale/angular-gettext/template-en.pot.
@@ -74,5 +89,9 @@ gulp.task('translations', function () {
         .pipe(gulp.dest(path.join(output_directory, 'i18n')));
 });
 
-// Gulp default task. Runs all other tasks before.
-gulp.task('default', ['js', 'css', 'fonts'], function() {});
+// Checks JavaScript using JSHint
+gulp.task( 'jshint', function() {
+    return gulp.src([ 'gulpfile.js', path.join( 'openslides', '*', 'static', '**', '*.js' ) ])
+        .pipe(jshint())
+        .pipe(jshint.reporter('default'));
+});
