+++ openslides/core/static/templates/index.html
@@ -8,7 +8,6 @@
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <link rel="stylesheet" href="static/css/openslides-libs.css">
 <link rel="stylesheet" href="static/css/app.css">
-<script src="static/js/openslides-libs.js"></script>
 
 <!-- Navbar -->
 <nav id="header" class="navbar navbar-inverse">
@@ -156,4 +155,4 @@
   </div><!--/.row-->
 </div><!--/#container-->
 
-<script src="/angular_js/site/"></script>
+<script src="openslides_javascript/site/"></script>
