+++ openslides/core/static/templates/projector.html
@@ -5,8 +5,6 @@
 <title>OpenSlides – Projector</title>
 <link rel="stylesheet" href="static/css/openslides-libs.css">
 <link rel="stylesheet" href="static/css/projector.css">
-<script src="static/js/openslides-libs.js"></script>
-
 <style type="text/css" ng-if="config('projector_backgroundcolor2')">
   #header {
     background-image: -moz-linear-gradient(top, {{ config('projector_backgroundcolor1') }}, {{ config('projector_backgroundcolor2') }});
@@ -29,7 +27,6 @@
   }
 </style>
 
-
 <div id="header">
   <img id="logo" src="/static/img/logo.png" alt="OpenSlides" />
   <div id="eventdata">
@@ -48,4 +45,4 @@
   </div>
 </div>
 
-<script src="/angular_js/projector/"></script>
+<script src="openslides_javascript/projector/"></script>
