+++ package.json
@@ -11,6 +11,7 @@
     "gulp-uglify": "~1.0.2",
     "main-bower-files": "~2.4.1",
     "yargs": "~1.3.3",
-    "po2json": "~0.3.2"
+    "po2json": "~0.3.2",
+    "gulp-jshint": "~1.11.2"
   }
 }
